Fix bug that prevents resume after pause
I noticed that resuming a paused strream doesn't work when I use my
headphones.

Turns out that in when you click the pause/play button in the iOS UI,
it emits Play and Pause events, but when you use the play/pause button
on a device, UIEventSubtypeRemoteControlTogglePlayPause is triggered
instead.

This commit fixes the bug that makes the pause button get stuck.