Speed up playback after play
Currently we let AVPlayer decide how much of the stream to buffer. In
practice it seems to take around 4-5 to hear something after pressing
play or switching channels. This feels like a long time, and doesn't
really match the experience in the browser, which starts playing
instantly.

This sets the buffer size to 1 second. It's still a little buffer to
deal with network issues, but it doesn't feel as sluggish.